fix(Exp_ZeroShot): patch 3 failing exps (21 bf16, 22 OOM, 23 GPT2 type) + log results 24-26
Fixes applied:
- exp_zs_21 TaskConditioningEntropy: BFloat16 unsupported for torch.log →
  cast logits to fp32 before softmax/log (numerical stability)
- exp_zs_22 ContrastiveHardNegatives: OOM 12.27GiB at bs=128 (2 vocab-logit tensors
  of 512×50265×fp32 each) → cut bs by 4x + empty_cache between forward passes
- exp_zs_23 KLDivergenceSignal: GPT-2 is causal LM loaded as MLM → switch to
  AutoModelForCausalLM + set pad_token = eos_token + shift logits [:, :-1] for
  correct next-token gather

Tracker results logged (successful 3/6 on first Kaggle H100 run):
- exp_zs_26 CodeAcrosticStructure: Droid T3 = 0.4383 (best of 21-26 suite!)
  but stability gap 10.12pt fails oral claim 3 (|Droid-CoDET| < 10)
- exp_zs_24 EntropyWatermarkDetection: Droid T3 = 0.3743, stability 3.47pt ✅
  but FAIL beat-Fast-DetectGPT (37.43 << 64.54)
- exp_zs_25 SyntacticPredictability: Droid T3 = 0.3628, stability 5.64pt ✅
  but HR < 0.95 on both benches

All 6 methods still below Fast-DetectGPT 64.54 baseline — failure mode consistent
with exp_zs_01-10. ZS suite as a whole cannot clear oral pass gate without fusion;
this motivates the NeurIPS claim 'no single signal dominates.'

Next: rerun 21/22/23 on Kaggle after fixes to populate full 21-26 row.

+++ Exp_ZeroShot/exp_zs_21_task_conditioning.py
@@ -122,13 +122,13 @@ def _task_conditioned_entropy(code: str, tokenizer, model, device: str = "cpu") 
     if device == "cuda":
         input_ids = input_ids.to("cuda")
 
-    # Get baseline logits
+    # Get baseline logits (cast to fp32 for numerical stability — log/softmax unstable in bf16)
     with torch.no_grad():
         outputs = model(input_ids)
-        logits = outputs.logits[0]  # (seq_len, vocab_size)
+        logits = outputs.logits[0].float()  # (seq_len, vocab_size), force fp32
         probs = torch.softmax(logits, dim=-1)
 
-    # Compute entropy
+    # Compute entropy in fp32
     entropy = -torch.sum(probs * torch.log(probs + 1e-10), dim=-1).mean()
     return float(entropy.cpu().numpy())
